put pest parser back into place
