+++ crates/hcl-edit/src/parser/error.rs
@@ -1,8 +1,16 @@
 use super::Input;
-use crate::parser::Location;
 use std::fmt;
 use winnow::error::{ContextError, FromExternalError, ParseError};
 use winnow::stream::{AsBytes, Offset};
+
+/// Represents a location in the parser input.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Location {
+    /// The one-based line number of the error.
+    pub line: usize,
+    /// The one-based column number of the error.
+    pub col: usize,
+}
 
 /// The result type used by this module.
 pub type ParseResult<T> = std::result::Result<T, Error>;
