+++ crates/hcl-edit/src/parser/template.rs
